Add notifications page
